Phase 1-4 optimizations: 3.01 → 7.2 tok/s (+140%), beats llama.cpp by 80%
Key optimizations:
- E8M0 LUT: replace ldexpf() with precomputed 256-entry float LUT (+128%)
- Serial router matmul: skip OpenMP for tiny 32×2880 gate GEMV (+3.6%)
- Float16 KV cache: uint16_t* with SIMD cvtph_ps reads, 384→192 MB (+1%)
- Fused RoPE + KV cache store: single pass instead of 3
- P-core thread pinning: SetProcessAffinityMask to avoid E-core stragglers (+3.4%)
- --quiet benchmark mode: buffer output during timing
- Persistent KV cache: incremental prefill across chat turns
- Software prefetching in matmul hot loops

+++ run_gptoss.h
@@ -130,9 +130,9 @@ typedef struct {
     // Output
     float* logits;      // (vocab_size,) final logits
 
-    // KV cache
-    float* key_cache;   // (num_layers, max_seq_len, kv_dim)
-    float* value_cache; // (num_layers, max_seq_len, kv_dim)
+    // KV cache — float16 to halve attention bandwidth (384 MB → 192 MB)
+    uint16_t* key_cache;   // (num_layers, max_seq_len, kv_dim) as float16
+    uint16_t* value_cache; // (num_layers, max_seq_len, kv_dim) as float16
 } RunState;
 
 // ----------------------------------------------------------------------------
